SCRD-5547 update_rpms.sh Cloud8 support/improvements
Make the appropriate changes to the update_rpm.sh script so that it
leverages the ARDANA_CLOUD_VERSION env var to determine which Cloud
version to build RPMs for.

Improved the determination of what git repo clones need RPMs to be
built for them, reporting meaningful reasons when skipping repos:
  * only use repos associated with the gerrit Ardana namespace.
  * only use repos checked out on a branch derived from the Cloud
    release we are building for, e.g. only master branch repos for
    Cloud9, or stable/pike branch repos for Cloud8.
  * check that both ardana-ansible and ardana are available and
    on the checked out on the required branch.
  * skip ardana-dev-tools

Also consolidated the determination and mapping of cloned repos to
associated IBS package, and ensure that we skip repos for which
there is no associated RPM being built as part of the standard
Cloud build.

Added initial support for the swiftlm project, though this will be
broken until we update the packaging process for this package to
consume the sources from gerrit directly, rather than using PyPI
as an intermediary step.

Avoid trying to do any RPM list retrieval or IBS operations if no
repos were detected as needing to be built as RPMs.

Update README.md section related to update_rpms.sh operation to
reflect the changes/enhancements made.

Change-Id: I7d52f0438e286e459031634cf5f766df26cae5a1