SCRD-5557 re-work vagrant ipv4/ipv6 settings management
Improve the management of the ipv4/ipv6 settings to integrate the env
vars associated with these settings into the top-level ardana global.

Update the ipv4/6 options setup such that we setup the appropriate
settings for V4 or V6 if the associated interface is enabled in the
env vars.

Add the --debug option which causes the vagrant helper to print out
the network settings being configured.

Change-Id: If1e3324b0c0b6f9fcd264c14ad5822b87b1a9352