SCRD-5556 Disable Spectre v2 (retpoline) mitigation (SCRD-5556)
For nested astack.sh based deployments within some OpenStack cloud
environments, i.e. when running astack.sh in a nested virtualisation
setup, our SLE 12 SP4 kernels, with Spectre v2 mitigation enabled,
fail to boot with a kernel panic. This change allows us todisable
those mitigation, by default, to avoid the problem, though the new
astack.sh option, --enable-spectrev2, will restore those mitigations
if desired.

This change also converts the SLE 12 SP3 & SP4 diskimage.json files
into templates, and adds support for the automatic templating of a
diskimage.json.j2 to create the required diskimage.json. This will
allow us to modify specific settings depending on the deployment
scenario if we need to. Note that we include either the template or
the static json in the manifest list indicating the inputs the image
build process, but not both. The generation of the list of files to
include in the manifest has also been simplified.

Update the scripts used to customise the packer image build process
to log their output under /root/ardana_image_build in the image. Also
cleanup these scripts somewhat and have the ardana.sh include the
build date in the customized motd file.

The patch also enhances the SLES image customisation scripts used by
packer to run zypper update to update the image using the locally
rsync mirrored zypper Updates (and Updates-test) repos appropriate
for the distro version if specified by packer via the kernel boot
parameters.

Add git ignorance of the generated diskimage.json files and links to
zypper repos we are serving to the packer image build process.

Eliminate some packages that we don't need to explicitly install
anymore, since we are not longer using it to bring up a build VM to
build python virtualenvs.

Add a root user definition to the SLES autoyast files to allow us to
re-enable the autoyast self update mechanism.

Change-Id: I53f9d94c3aec9f687b2f62492031196a76ceae5a