SCRD-6805 Update Cloud9 default milestone to 9
Only cobbler testing needs to download the ISO, so we haven't needed
to exercise this code except for cobbler testing, since the SCRD-5471
changes landed. However the previous default milestone, M4, seems to
have been removed so we fail when attempting to download the ISO.

This patch updates the default milestone to 9.

Change-Id: I9d73f2420c3e5319983f7b0026f261c82f52feab