SCRD-5899 run-cobbler testing enhancements (bsc#1110061)
Issues: SCRD-5899

In order to ensure that we are validating the wipe_disks.yml issue
reported in bsc#1110061 we need to run wipe_disks.yml in a mode
similar to what the customer experience would be, i.e. remove all
but the localhost entries from the deployer account's known_hosts
file before running the wipe_disks.yml playbook.

To support this the verify-bm-install script that, that oul cobbler
testing process uses to validate the re-imaging process did actually
re-image the specified nodes, has been refactored as follows:
  * dynamically generated verify-bm.sh script split out into a new
    script under bin/deployer.
  * restructure verify-bm-install script to make use of the
    run-in-deployer.sh tooling to run scripts on the deployer.
  * add support to verify-bm-install to trigger running the
    wipe_disks.yml playbook unless --no-wipe-disks option is
    specified.
  * add support to verify-bm.sh to enable running the wipe_disks.yml
    playbook if the --wipe-disks.yml option is specified.

Depends-On: Iee57599fdff1d0fd219fa90f1de2a913ecd9fe8e
Depends-On: I5922792fc3f435810933c1bc1741a0494b4e4575
Change-Id: Iad053df8a48f6a18fa0be535f436b8a4bea00cfd