SCRD-7427 packer image shouldn't have persistent net rules entries
The cleanup.sh used for SLE 12 image creation was incorrectly deleting
the script that generates the persistent net rules rather than the
generated persistent net rules.

Also cleaned up the list of files passed to the md5sum calculation so
that it ignores the dynamically generated Vagrantfile, and includes
the shell scripts used to create the image.

Change-Id: Ie2714c3782e30f3a4fb9bc7515458d6bdacba37a