SCRD-7471 Fix packer image build for ansible 2.4 (SCRD-7471)
The logic used to test the registered result of running md5sum against
the input files used to generate the packer image was broken under
Ansible 1.9, but worked due to bugs in how that version of Ansible
was implemented.

This patch fixes the logic to test for the registered result being
changed rather than failed.

Also updates to use the latest version of packer, 1.3.4.

Change-Id: I175ca632e955ec36c9c78a23863b98f9a94c8a1b