SCRD-8297 astack.sh needs to use SLE updates-test repos (SCRD-8297)
Due to recent changes in Cloud9 Devel and Staging package sets, we need
to pull in the SLE12 SP4 Updates-test repos when deploying clouds based
on the Devel or Staging repos. Cloud9 Pool and Updates based deployments
will still use just the SLE Updates and Pool repos for now.

Change the default set of repos to include updates-test.

Fix handling of the updates-test repo specification to work correctly
when mirroring and setting up the repos on the cloud nodes.

Increase the retry delay for rsync failures when mirroring repos to 30.

Change-Id: I320055d9135a8c7f5548a76439fa7421ebda2022