SCRD-8390 update-ansible.sh enhancements (SCRD-8390)
Sometimes you just want to quickly test a change to the ansible code in
an Ardana ansible repo, e.g. neutron-ansible, without having to commit
it to the openstack/ardana/ansible repo and re-run the configuration
and then ready-deployment.yml playbook. This patch adds a new --scratch
option to facilitate this which will rsync the specified Ardana ansible
repo contents to the scratch area, rather then the openstack area,
and run the specified playbook if any after doing so.

Improvements to the rsync exclusion list used when rsync'ing changes to
the deployer node, avoiding sync'ing unnecessary files.

Also enhanced the run-in-deloyer.sh script's ssh command to attempt to
create a psuedo terminal session so that ansible command output will be
colourised when running commands like update-ansible.sh interactively.

Change-Id: I61e7797f3ac30f04561477e5e560df226fb39a4d