SCRD-7471 Add --c9-milestone option (SCRD-7471)
Add support to astack.sh for the --c9-milestone option that can be
used to specify which milestone's ISO to download and use as the
source for Cloud9 packages.

This revises/updates the purely environment variable based mechanism
that previously existed, and updates the default milestone to GMC1.

Updated README.md to add some details about the new option.

Also updated some text in the README.md to reflect changes that have
gone into ardana-dev-tools over the last few months. And cleaned up
a few typos.

Change-Id: I0e39305c6765c9a9fd6ff2d818663c2f76d5dc01