SCRD-9090 Misc fixes for cleanup-slave (SCRD-9090)
Ensure that vagrant is setup appropriately before trying to use it,
by running the dev-env-setup script with the --vagrant-only option.

Use correct name for the .astack_env file in cleanup-slave so that
any such files get cleaned up when the script is run. Also correct
the name in other places, mostly comments and documentation.

Fix determination of ISO link targets to not fail if .cache-ardana
doesn't exist.

Change-Id: Ief4e5a6a0df6b5e62c8c4630d34029315490f8f7