SCRD-9076 adt model rhel support (SCRD-9076)
Renaming the role of the second node from COMPUTE to RESOURCE meant
that we didn't trigger the code in the Vagrant helper to select the
RHEL vagrant box when bring up the nodes.

Also complete the renaming process by specifying appropriate values
for the resource cluster and the node prefix.

Change-Id: I70b63dc67e4f76329cc3fb7f3de4f5866751fbc7