SCRD-9093 Default to Vagrant 1.8.7 (SCRD-9093)
We have been defaulting to running Vagrant 1.7 stream for a long
time to maintain compatibility with the legacy HOS Zuul CI testing
workflow but that is no longer a requirement so we can move the
defaulty Vagrant version forward to something newer.

It would be nice to move forward to using the latest version of
Vagrant, e.g. 2.2.4, as our default, but we have seen issues with
the network setup for RHEL/CentOS compute nodes in versions of
Vagrant newer than 1.8.7, so we will settle moving to 1.8.7 for
now.

Also remove the no longer required rack, fog and openstack plugins
from the list of 1.7 stream Vagrant plugins since we have moved
forward to using latest vagrant-libvirt plugin version in that
stream for some time now. And since that means that all streams
now use a common set of plugins, we can consolitate the plugin
list into a single common list, referenced by the stream specific
settings.

Also add support for the Vagrant 2.2 stream.

Change-Id: I24b79f388c1715329fd206391ea93d7c4e4c1440
(cherry picked from commit 8ac953433aa08a1202fafcb453568c00dfea17c2)