Enable Cloud8 deployments from master branch (SCRD-9077)
The way that input models were handled in the past by the Vagrant
helper and staging code caused problems when trying to deploy Cloud8
using the master branch. However the new input model management
scheme doesn't suffer from these limitations.

This patch updates the setup-vagrant-input-model tool so that it
defaults to the appropriate branch, if no branch is specified on
the command line, based upon the environment settings that will
exist when run by astack.sh run.

Change-Id: Ia97d766906af10b072228c0f1fd3451be24e69e6