Increase adt controller disk to 150GiB (SCRD-9077)
The initial disk provisioning only grows the root partition to 20% of
the capacity, which is 20GiB, and that is not sufficient for Cloud8
deployments where we run out of disk space rsync'ing the SLE 12 SP3
repos to the controller node.

Increasing the size to 150GiB, like we already use in the other input
models avoids this problem.

Change-Id: I44829ae5338af022d106d2411f0d514d86245c07