Install relevant packages to support osc command (SOC-7471)
Recent versions of the osc package from OBS don't trigger the install
of all the packages that need to be installed to support running a
build of packages locally so we need to explicitly list some of the
dependent packages. Additionally we have to deal with that fact that
some of the base package generic names have '_' in them which needs
to be remapped to '-' because '_' is a reserved character in Debian
package names.

Change-Id: Ifc7eec85865363a349a36044e3bfe3a06279b9ca