Add PolicyKit rule for unprivileged access (SOC-6795)
The libvirt tools allow for authentication via PolicyKit. This gives
unprivileged users access to libvirt without the need to add the user
to any system groups.

Currently the ardana-dev-tools require a user to re-login after the
dev-env-setup script has been run. This is due to the unprivileged user
being added to the kvm and libvirt groups, but the groups are not
active for the user until re-login.

With this PolicyKit rule, the unprivileged user has access to libvirt
without the need to re-login. This allows dev-env-setup utility to
proceed without interruption.

These changes do not remove the addition of the user to the kvm or
libvirt groups. The changes do remove the validation of the user in the
groups during the environment setup.

Signed-off-by: John L. Jolly <jjolly@suse.com>
Change-Id: I521adb31158309339e02e4f29b5c15487ddf4736