Use appopriate hosts lists in top level playbooks (SOC-9077)
Instead of using the 'all' target we should use the appropriate hosts
specification, e.g. cloud, deployer or localhost, in the top level
playbooks.

Only generate a 'deployer' hosts group if there isn't a node with that
name in the set of cloud VMs. Ensure we upload generated SSH key pair
to the deployer node, whether it is the member of the deployer group,
or called deployer.

The cloud-setup role no longer depends on the vagrant-vm role, so we
remove that role from the meta dependencies list.

Speed up the cobbler-vm role by parsing the list of vms out of the
astack-ssh-config rather than from the output of the vagrant status
command.

Change-Id: I4e11a3d38f65c08f644f65c369285d72ea122004