Add user to sudoer file (SOC-6795)
Early in the astack.sh process the root password is required in order to
perform system installation tasks. Eventually the authorization
times-out and cannot be refreshed by Ansible interactively. This is
resolved by adding the user to the sudoer file of the system. Before
this change the instructions required the manual addition of the user to
the sudoer file.

This change adds to the ardana-env script functionality that will add
the user to the sudoer file.

Signed-off-by: John L. Jolly <jjolly@suse.com>

Change-Id: If56c61f32524b8dc3d33d893720cb87729dd8a5b