Re-add python-xml and iputils to Vagrant box images (SOC-9077)
A previous change incorrectly trimmed the python-xml package from the
list of packages to install in the SLE 12 SP4 Vagrant box image, which
means we can't use the Ansible zypper module to install packages.

Also re-add the iputils package to the SLE 12 SP3 and SP4 Vagrant box
images so that ping can be used without needing to install additional
packages; will be installed by the SOC/CLM deployment anyway.

Change-Id: I75a924ff30106dfe1eca1b6a55a1ef6793a0fbf3