Remove support for proxy cache (SOC-9077)
The proxy cache was originally intended as an alternative to having
to download/mirror the SLE & SOC repos to your local machine before
deploying the cloud. However those repos can be updated while the
cloud is being deployed which can lead to instability or packages
having different versions on nodes within the cloud dependending on
when the repo package manifests are sampled.

For a long time now we have only been using the mirroring approach
without significant issues, so we can safely remove the proxy cache
support.

Change-Id: I1b037f2154ce50a817db5d99404c2ead4226b0b0