Add support for update and upgrade testing (SOC-9742)
Add tooling to implement Cloud update testing, e.g. updating to latest
Updates packages, or testing Updates-test packages against a cloud with
Pool+Updates, or testing update from Devel to Staging.

Add tooling to implement Cloud upgrade testing, e.g. upgrading from
Cloud8 to Cloud9 for devel, staging, updates or updates-test.

Add support to both run-update and run-upgrade to allow individual
service components or groups of them to be enabled/disabled in the
initial cloud deployment.

Add a script to apply any mandatory Cloud9 input model deprecations.

Add a script that can be used to migrate a Cloud8 system to use fernet
token provider which potentially needs to be done before we start the
upgrade.

Add a hack to to deploy-cloud.sh script to install python-oslo.log to
workaround SOC-10253 which occurs when freezer-agent or monasca-agent
aren't being installed.

Add new input models to the available set of vagrant models, which
are tailored to support Cloud8 to Cloud9 upgrade testing, leveraging
the new ardana-dev-tools based input-model management mechanism. The
new models are:
  * std-upgrade - similar to a mid-scale model, tailored to exercise
    separation of services into dedicated control plane clusters.
  * std-basic - a cutdown version std-upgrade, that drops the MML
    stack services and combines the DBMQ & Swift clusters to allow
    us to exercise the basic OpenStack services without MML stack
    overheads.

Convert std-split model into an ardana-dev-tools managed input model,
increasing the size of the MML cluster nodes to 10G RAM, the deployer
node size to 3G, and the DB nodes to 3G to support testing Cloud 8 to
9 upgrade without hitting OOM issues.

Change-Id: Ib4e3972747da5cdb30843c38fde94afc76256371