Eliminate dependency of Vagrant ssh private keys (SOC-9077)
Once the cloud-vm-setup.yml play has configured SSH key access within
the cloud nodes, we no longer need to rely on the use of the Vagrant
private keys when ssh'ing into the cloud nodes.

The new setup-cloud-inventory.yml playbook, and the associated utility
script setup-cloud-inventory, will trigger updating the host_vars files
and astack-ssh-config files to use the correct key files.

We add an appropriate default ansible_ssh_private_key_file setting in
the group_vars/all/ssh file which will be used if no host specific
override is set.

As a result of this it should now be possible to complete deploying
a cobbler-ised cloud.

Fix error in comments in the setup-vagrant-inventory script.

Change-Id: I14ab0e1fa74f705bf3ee9dc90a3b24f7685b5c6b