Allow .astack_env settings to be overriden (SOC-9742)
Modify the generation of the .astack.env file to allow the specified
settings to be overridden on the command line via environment vars.

Also limit the set of saved environment variables to those used by
the ardana-dev-tools tooling to avoid undesirable interactions due
to stale or inappropriate environment variable values.

Change-Id: I42ae6677fe438b12faa0b814ce781f2696bfcdde