Add Cloud tags to relevant repo metadata (SOC-9742)
The Ardana update workflow expects that specific Product and Updates
tags will be advertised by the registered system repos, or it will
fail with an error naming the required repo tags that were not found.

Since the Devel and Staging built media don't include these tags out
of the box, if we want to exercise the Ardana update workflow on a
cloud installed using Devel or Staging package sources, we need to
explicitly add these tags to the metadata of the repos once they have
been uploaded to the deployer.

Change-Id: I1ba3728577110efb20682d816c0c3b40c74dcf57