Add playbooks to manage rebooting cloud nodes (SOC-9742)
Add the reboot-node role, and associated playbooks which can be used
to reboot the entire cloud, just the deployer, or all nodes but the
deployer. By default multi-node reboot operations will run with a
serial setting of 1, so that only one nodes is rebooted at a time.
This can be overridden by specifying the desired serial value using
the reboot_serial variable on the command line.

Change-Id: I0138bdab3132a22e536132e3f13d8d962842e80a