Update to Ansible 2.5.15 (SOC-7471)
Need to ensure that we install/upgrade pip, wheel and setuptools to
latest versions to avoid issues with dynamic importing of the Python
pkg_resources module.

The Ansible 2.5 stream version of the synchronize modules is broken
in the same was as the Ansible 2.4 stream when it comes to delegated
operation.

Address some deprecation warnings related to the use of the search
test being used as a filter. Also switched to using '~' in jinja2
expressions for string concatenation rather than '+' as it avoids
potential type promotion issues.

Enable Ansible logging to the logs/vm_host_ansible.log file as part
of the ardana-env command; this ensures we capture logging for all
ansible-playbook actions, including the initial dev-env-setup.

Change-Id: I2b8f3631a4cfa49e5d690fde89e311e4f852ffba