Update to Ansible 2.7.12 (SOC-7471)
Switch to Ansible 2.7.12 for running the ardana-dev-tools tool chain.

The Ansible 2.7 stream version of the synchronize modules is broken
in the same was as the Ansible 2.4, 2.5 and 2.6 streams when it comes
to delegated operation.

Address the deprecations for the apt and zypper actions; they now can
accept a list of packages to be installed directly, rather than relying
on hackery to optimise these actions being run looping over a list of
packages.

Additionally the zypper action now supports being able to pass the
necessary pre-command args to allow us to install an RPM from a file
rather than from a repo.

Simplified the dev-env-radvd role's install of the radvd package to
use the package action, which automatically detects the appropriate
system specific package manager to use.

Change-Id: I733c09d22fa9f36b374feed0d7b7c6c6dc16e2fa