Align repo names/paths with product expectations (SOC-10033)
The SOC product expects slightly different names for locally/manually
managed SLE and SOC repos, so update the astack.sh repo definitions
and management code to align with the product expectations.

Also update the product/updates tag management, used to make the SOC
deployment think a Devel or Staging code base is a Product and Updates
repo, to allow the key tag values to be specified with the repo specs.
This allows us to correctly handle HOS8 tags as well as SOC8/9 tags.

Finally ensure that appropriate defaults are setup when specifying the
--c8-hos option.

Change-Id: Ie21c0b113c293401cd29210a1430ded6437e53b7