Correctly implement retries for get_url operations (SOC-7471)
When retrieving artifacts from the artifacts servers, or from the
clouddata mirrors, using get_url we should use a repeat...until
construct to ensure that we retry a reasonable amount of times
with appropriate delays, before giving up.

In the case of artifacts served by the legacy CI artifacts servers
these should never by changing once published, so we really just
want to retry in the event of a network glitch, delaying for just
1 second between retries.

In the case of artifacts served by the clouddata mirrors, we need
delay sufficiently to allow for delays when swapping in an updated
ISO. Also needed to add in the missing register and until directives
to enable the repeat..until handling.

Change-Id: Ifd02569e6913cd2adc933495f934c0d7f7bf0fe4