Support RHEL/CentOS computes for Cloud 8 and 9 (SOC-10966)
Enhance master branch support for Cloud8 deployment to enable support
for RHEL/CentOS compute deployments. Previously only Cloud9 compute
deployments were supported. To achieve this we rename the existing
Cloud9 support role to cloud9-rhel-deployer, and copy over the Cloud8
version of the role from the stable/pike branch as cloud9-rhel-deployer
and select which version of the role to use based on Cloud version we
are deploying.

Additionally fixed the get-ardana-artifacts.yml triggered actions to
correctly download the ISOs if we are enabling RHEL compute support as
we use the RHEL/CentOS ISO to provide the relevant repos for that type
of compute node.

Also fixed the lib/ardana_vagrant_helper.rb to use the cloud version
rather than the active gerrit base branch to determine where to find
the relevant ISO files. Updated the error messages generated from a
failed attempt to attach an ISO to report the path being used as this
aids in debugging such issues.

Change-Id: I728c6211a0d5a66386997623115123d23ac581bd