Use pre-defined SHA256 values for SLE ISOs (SOC-9742)
Dynamically retrieved SHA256 values are only really required when the
target ISO is subject to rebuilding, which is not the case ones the
SLE version has been released and we have switched to using GM ISOs.

This also avoids issues when working with download.suse.de as a source
for ISO images, as is the case for SLE 12 SP3, where no associated file
holding a SHA256 is available.

Change-Id: I01bcb24b76cf11de090f8d3e7db66ccac9972457