Remove Cloud8 RPM build dependecy (SOC-7471)
Previously the Cloud8 RPM build process needed the ardana.git repo
cloned beside ardana-dev-tools, in addition to the ardana-ansible.git
repo, to be able to build the ardana-ansible RPM package. This is
no longer the case so we remove the special handling in the master
branch version of update_rpms.sh for this scenario.

Change-Id: Idcf0270b51e7735f18d8cf3ef45aad01f41dcbc6