Refresh versions for tools we use (SOC-7471)
Update default Ansible version to 2.7.15. Note that 2.7.16 is available
upstream, but the latest version available in the local bandersnatch
mirror used by the legacy CI is 2.7.15.

Tested support for running against Vagrant 2.2.7 if specified via
ARDANA_VAGRANT_VERSION.

Update to using Packer 1.4.5. Note that 1.5.1 is the latest available
but it doesn't work out of the box with our packer templates.

Add support to bin/dev-env-setup for
  * openSUSE Leap 15.1.
  * Ubuntu 19.10 (Eoan)

Re-work the libvirt dependent package install process, updating the
required package lists for existing and new Ubuntu versions.

Workaround Ubuntu Xenial python-setuptools issues when creating the
ansible venv by limiting the version installed by pip to be < 45.

Fix vconfig output checking for non-existent devices to handle the
fact that eoan output is slightly different when setting up the
routing between the host and the deployer.

Change-Id: Ie04eb300e719ae14f87060574edaea7323def3e6