ardana-dev-tools needs gawk installed (SOC-7471)
The input-model-inventory role tasks use the GNU awk gensub() function
so we need to ensure we have gawk installed, and that we explicitly use
gawk when performing that action.

To ensure we are doubly covered, we add the gawk package to the list of
packages that the ardana-env script attempts to install, and to the set
of packages that are installed by the dev-env-install.yml playbooks.

Change-Id: I3da6099f7a21e10b68b15e108f7e54da3afbe6e6