Add std-upgrade-min model (SOC-9742)
Add a new std-upgrade-min model based on the std-upgrade model that
combines the swift services with the DB & MQ services, similarly to
the std-basic model, to minimise VM footprint for testing within a
128GB memory model reliably.

The std-upgrade model, while representing only about ~100GB VM load,
leads to swapping on a 128GB system because it appears that KSM is
not detecting any shareable pages between the various VM instances.
Combined with the I/O activity on the system leading to buffer cache
load, we exceed the capabilities of a 128GB model.

The std-upgrade-min reduces the memory footprint by ~15GB and the
number of VMs by 4 and seems to avoid hitting the swapping issues.

Change-Id: Ie91f36cefd882b030dc57e106717e3f83ea70c82