Workaround osc broken Ubuntu packaging issues (SOC-7471)
Recent upstream changes to the Debian packaging used for building the
osc command have switched to using Python3 based setup requiring the
M3Crypto Python module to be available for Python3. However only the
Eoan (19.10) and Focal (20.04) release streams currently have the
python3-m2crypto package available meaning that the latest upstream
builds of the osc package for Xenial (16.04) and Bionic (18.04) are
broken and can't be installed.

For existing Ubuntu based DevEnv deployments, where osc is already
installed, there shouldn't be a problem, for now, as the broken
packaging means that the latest version can't be installed and thus
the previous Python2 based version persists and works correctly.

However for fresh Ubuntu based DevEnv deployments, the upstream
osc command can't be installed; to workaround this I have created
http://download.opensuse.org/repositories/openSUSE:/Tools/ with
Bionic (18.04) and Xenial (16.04) builds of the osc package using
the Debian packaging as it was just prior to the Python3 switch.
The version has .1 appended to it, which is sufficient to ensure
that it will be preferred over the existing upstream version,
until they release an update for it.

This workaround repo will be automatically included on Bionic and
Xenial systems as part of the fresh deployment process when the
ardana-env script is run on a relevant Ubuntu system; if you want
it to fix up an existing system you need to uninstall osc and remove
the existing /etc/apt/sources.list.d/osc.list file, and then run
ardana-env again.

Change-Id: I96649a6155ed522597c7dead61a93eeeb9152201