Tweak RabbitMQ settings for virtual testing (SOC-11083)
Add a new script - virtual-testing-rabbitmq-tweaks - that is intended
to be run on a deployer to tweak the rabbitmq default settings to be
better suited for running in a virtualised testing environment.

Updated astack.sh to call virtual-testing-rabbitmq-tweaks during the
initial deployer node setup, after copying up the input model.

Also VM settings for the split and upgrade testing models:
  * Use 8GB and 4 vCPUs for nodes providing RabbitMQ services
  * Use 16GB and 4 vCPUs for nodes providing MML services
  * Use 16GB and 4 vCPUs for nodes providing Core openstack services
dac-3cp model change to:
  * use 35GB and 4 vCPUS for controllers
  * use 2 vCPUs for the compute

Change-Id: Ie105d7e4c44e816c10e89b9e566e652e6b219c59