Make Vagrant disk cache settings configurable (SOC-7471)
Add support for being able to configure the Vagrant disk cache settings
for the boot and extra disks that are attached to the Vagrant managed
VMs.

Also switch from defaulting to "unsafe" to "none" for the boot volumes
and from "default" (aka "writeback") to "none" for the extra disks. The
"none" mode equates to Direct I/O access to the libvirt backing volumes
in the host environment which should eliminate some host memory pressure
issues when running large numbers of VMs on the same host.

The newly added ARDANA_DISK_CACHE env var can be used to change the
default caching mode ('none') to another value. Similarly a 'cache'
value can be specified in the 'ci_settings' block in the servers.yml
file in the 'boot' and 'extras' sections under 'hardware'->'disks'.

Change-Id: I4f7b682adb90a265dc7854b9a7e41ea5fb015b04