Add plumbing for SLE12 SP3 LTSS repos to Cloud8 deployments (SOC-11222)
Add the SLE 12 SP3 LTSS repos to the list of repos that we need to
mirror for Cloud8 deployments.

Update the cloud-setup repo list to include the SLE 12 SP3 LTSS repos
so that they will be configured if enabled in Cloud8 deployments.

When deploying Cloud8, ensure that the LTSS repos will be included
in list of SLE repos that we setup on the cloud nodes.

Change-Id: I68677401f82457da02e517f0e1ebae00ba7882bf