Add plumbing for SLE12 SP3 LTSS repos to Cloud8 deployments (SOC-11222)
Add the SLE 12 SP3 LTSS Updates repo to the list of repos that we need
to mirror for Cloud8 deployments.

Update the c8setup repo list to include the SLE 12 SP3 LTSS Updates
repo so that it will be configured for Cloud8 deployments.

When deploying Cloud8, ensure that the LTSS Updates repos is
included in list of SLE repos that we setup on the cloud nodes.

(cherry-picked from commit b84db231444816ff9e9bbc9d6a1b38355bb60e71)
Change-Id: I68677401f82457da02e517f0e1ebae00ba7882bf

Change-Id: I223a60d6a177cfb530f3c7c7e3745cbe1bbcdaa2