Move dev-env settings to group_vars/all (SOC-7471)
This was never done as part of the previous similarly re-structuring
to move all settings into the group_vars/all area so just getting it
done now.

Also cleanup some meta dependencies for dev-env that don't make sense
now that we have moved relevant settings out of role specific defaults
and in to the group_vars/all area.

Change-Id: I138e01663fd0814721ad233e8c2872d2315a8d05