Disable sync of ansible sources to deployer (SOC-7471)
Disable the automatic sync of local ansible sources to the deployer node
when bringing up the deployer node. This feature can still be enabled by
specifying the --ansible-sync option if desired. The --no-ansible-sync
option can still be used to negate a previously specified --ansible-sync
option.

Enhance sync-ansible-sources role to only sync sources if local the repo
is cloned on the appropriate branch/stream, avoiding mistakenly sync'ing
Cloud 9 ansible sources to a Cloud8 deployer or vice versa.

Update run-upgrade to remove the no longer needed --no-ansible-sync.

Remove role name in name: lines in the sync-ansible-sources role's task
files as this is no longer needed in modern ansible versions due to the
role name being a predix of the reported TASK lines.

Change-Id: Icb857b0b2469a21d2c18a963c9254cba8b772a1e