Improve overrides repo handling (SOC-7471)
Fix the "alias" of the Cx-Overrides repo to be the same as the "name"
when it is defined. This avoids using an alias value that includes
spaces.

Update the run-upgrade script so that it ensures that the overrides
repo is setup on all nodes, not just on the deployer node; this will
match how the overrides repo is setup during a normal deploy.

Change-Id: I232387192dd20cf947d8d4e547e5f0c5416e3d9e