Move the running of createrepo to the deployer (SOC-7471)
The createrepo command is Python2 based and not available on recent
Python3 based distro releases. Moving the running of the createrepo
command to the deployer avoids issues in such cases as createrepo
is available in SLE 12 SP3/4.

At the time when we need to run createrepo on the deployer, we will
not yet have setup any repos from which to install the createrepo
package, so we need to include createrepo in the list of packages
that are pre-installed in the Vagrant box image.

Change-Id: Iaf07ca0f6a477b29c1eb49142410d9ca16313ee2