Remove Python sh module from ardana-env (SOC-7471)
A long time ago I added sh to the list of Python modules that need to
be available either in the ansible virtualenv or on the host, as part
of setting up the environment for running the Ardana Dev Env.

However I never actually added any tools that required the sh module.

So we can safely remove it, and simplify the ardana-env setup a bit.

Change-Id: I7dd1ffe2b241552c30d38cc1c83f729ee13f4b94