osc --version hangs for up to 25 secs doing DBUS check (SOC-7471)
Noticed the osc --version command taking a long time when running in
a VNC session but not in an SSH session. Tracked this back to having
DBUS_SESSION_BUS_ADDRESS env var setup under VNC session, but not in
the SSH session.

Unsetting the env var when calling osc --version speeds up ardana-env
significantly in such a scenario, and when just getting the version
we don't care about DBUS integration anyway.

Change-Id: Ie7d7bf0b1407739c24773a68667fce178beff92a