Fix buildservice repo refresh handling (SOC-7471)
When trying to use the --obs-repo option to specify an OBS repos for
testing purposes I ran into issues with the appropriate GPG signing
key not being auto imported by the refresh command.

This sounded familiar, and it turns out that when I previously fixed
this for the deployer and general cloud repo setup I failed to apply
the same fix to the suse-buildservice-repos role, namely to pull in
the working zypper_repository task list from the cloud-refresh-repos
role.

Change-Id: I8d5108740a3b41766cb309517d7b12f92aaafbd0