Fix incorrectly constructed Vagrant download URL (SOC-7471)
Previous fix mistakenly included the vagrant version in the download
URL as had previously required for upstream downloads, and only fixed
this issue for the SHA256 URL construction.

Also add support for Vagrant 2.2.17 and 2.2.18, though 2.2.17 seems to
be broken due to being based on Ruby 3.0.1, which causes problems for
plugin gem dependencies.

Change-Id: Ibbd3a6c720813ad4f4136e22ad3ebdd89045bf72