Use ardana_family instead of ansible_os_family (SOC-7471)
In some cases the ansible_os_family fact will not be set by the Ansible
gather_facts (setup) module run. To workaround this we can update the
definition of ardana_family to still get a meaningful value in such
cases, and use ardana_family instead of ansible_os_family to determine
which distribution is running on the node.

This specific issue appears to affect Ubuntu 20.04.3, where the system
default Python has been switched to Python3, and the Ubuntu standard
ansible package (version 2.9.6) has been installed.

Since we only support Ubuntu and openSUSE Leap installations, and lsb
is usually installed on a Ubuntu distro, this workaround is viable for
now.

Change-Id: Iaeae6f2cc7dcdaa45a5248d1b592c0c697297c0e